+++ src/base/PluginTrait.php
@@ -6,6 +6,7 @@ use verbb\giftvoucher\services\PdfService;
 use verbb\giftvoucher\services\RedemptionsService;
 use verbb\giftvoucher\services\VouchersService;
 use verbb\giftvoucher\services\VoucherTypesService;
+use verbb\giftvoucher\storage\CodeStorageInterface;
 
 trait PluginTrait
 {
@@ -68,14 +69,31 @@ trait PluginTrait
         return $this->get('voucherTypes');
     }
 
+    /**
+     * Get the code Storage
+     *
+     * @return \verbb\giftvoucher\storage\CodeStorageInterface
+     *
+     * @author Robin Schambach
+     * @since  2.0.16
+     */
+    public function getCodeStorage(): CodeStorageInterface
+    {
+        return $this->get('codeStorage');
+    }
+
     private function _setPluginComponents()
     {
+        /** @var \verbb\giftvoucher\models\Settings $settings */
+        $settings = $this->getSettings();
+
         $this->setComponents([
             'codes' => CodesService::class,
             'pdf' => PdfService::class,
             'redemptions' => RedemptionsService::class,
             'vouchers' => VouchersService::class,
             'voucherTypes' => VoucherTypesService::class,
+            'codeStorage' => $settings->codeStorage,
         ]);
     }